refactor: Move row as child of table body
